Move vehicle service to @Common
