+++ projects/common/src/models/vehicle/index.ts
@@ -1,0 +1,5 @@
+export * from './vehicle-item.model';
+export * from './maintenance-item-create.model';
+export * from './maintenance-item-response.model';
+export * from './maintenance-item-update.model';
+export * from './vehicle-create.model';
