+++ projects/common/src/models/vehicle/search-vehicle-response.model.ts
