+++ projects/common/src/models/vehicle/supplier-search-response-model.ts
