+++ projects/common/src/models/vehicle/vehicle-create.model.ts
