+++ projects/common/src/models/vehicle/vehicle-item.model.ts
