+++ projects/common/src/models/vehicle/vehicle-params.model.ts
