+++ projects/common/src/models/vehicle/vehicle-update.model.ts
