+++ projects/common/src/services/index.ts
@@ -1,4 +1,5 @@
 export * from './nav-bar.service';
 export * from './auth/auth.service';
 export * from './town/town.service';
+export * from './vehicle/vehicle.service';
 export * from '../../../order/src/lib/utils/order-list-utils';
