+++ projects/common/src/services/vehicle/index.ts
@@ -1,0 +1,1 @@
+export * from './vehicle.service';
