+++ projects/vehicle/src/lib/components/maintenance-history/maintenance-history.component.ts
@@ -1,3 +1,4 @@
+import { VehicleService, VehicleListItem } from '@Common';
 import { downloadFileFromResponse } from '@Common';
 import {
   TitleComponent,
@@ -13,14 +14,12 @@ import { Component, OnInit } from '@angular/core';
 import { MatTabsModule } from '@angular/material/tabs';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { VehicleListItem } from '../../models/vehicle-item.model';
-import { VehicleService } from '../../services/vehicle.service';
 import { MaintenanceListComponent } from '../maintenance-list/maintenance-list.component';
 import { MaintenancePlanListComponent } from '../maintenance-plan-list/maintenance-plan-list.component';
 import { MaintenanceRepairListComponent } from '../maintenance-repair-list/maintenance-repair-list.component';
 
 @Component({
-  selector: 'lib-maintenance-history',
+  selector: 'mp-maintenance-history',
   standalone: true,
   imports: [
     TitleComponent,
