+++ projects/vehicle/src/lib/components/perform-maintenance-plan/perform-maintenance-plan.component.spec.ts
@@ -1,3 +1,4 @@
+import { VehicleService } from '@Common';
 import { LateralDrawerService } from '@Common-UI';
 
 import { Location } from '@angular/common';
@@ -13,8 +14,7 @@ import { ActivatedRoute } from '@angular/router';
 import { of, throwError } from 'rxjs';
 
 import { PerformMaintenancePlanComponent } from './perform-maintenance-plan.component';
-import { SupplierSearchResult } from '../../models/supplier-search-response-model';
-import { VehicleService } from '../../services/vehicle.service';
+import { SupplierSearchResult } from '../../../../../common/src/models/vehicle/supplier-search-response-model';
 
 describe('PerformMaintenancePlanComponent', () => {
   let component: PerformMaintenancePlanComponent;
