+++ projects/vehicle/src/lib/components/perform-maintenance-plan/perform-maintenance-plan.component.ts
@@ -1,3 +1,4 @@
+import { VehicleService } from '@Common';
 import {
   BackArrowComponent,
   TitleComponent,
@@ -38,9 +39,8 @@ import {
   finalize,
 } from 'rxjs/operators';
 
-import { MaintenancePerformRequest } from '../../models/maintenance-perform.model';
-import { SupplierSearchResult } from '../../models/supplier-search-response-model';
-import { VehicleService } from '../../services/vehicle.service';
+import { MaintenancePerformRequest } from '../../../../../common/src/models/vehicle/maintenance-perform.model';
+import { SupplierSearchResult } from '../../../../../common/src/models/vehicle/supplier-search-response-model';
 import { CreateUpdateSupplierServiceDrawerComponent } from '../create-update-supplier-service-drawer/create-update-supplier-service-drawer.component';
 
 @Component({
