+++ projects/vehicle/src/public-api.ts
@@ -2,6 +2,6 @@
  * Public API Surface of vehicle
  */
 
-export * from './lib/services/vehicle.service';
 export * from './lib/components/vehicle-list/vehicle-list.component';
+export * from '../../common/src/models/vehicle/index';
 export * from './lib/vehicle-routing.module';
